Improved stair placement
When adding up/down stairs to dungeons, it now tries first to place doors in
the middle of any room, scanning each room for a centre point that is
surrounded by at least one ring of empty spaces.  If that fails, than it
reverts to picking a random offset and searching for the first available empty
spot from there.